+++ src/plugins/_generated/metas.ts
@@ -7,8 +7,6 @@ import { META as accountingMeta } from "../accounting/meta";
 import { META as canvasMeta } from "../canvas/meta";
 import { META as chartMeta } from "../chart/meta";
 import { META as editImagesMeta } from "../editImages/meta";
-import { META as encoreDefineEncoreMeta } from "../encore/defineEncoreMeta";
-import { META as encoreManageEncoreMeta } from "../encore/manageEncoreMeta";
 import { META as generateImageMeta } from "../generateImage/meta";
 import { META as manageSkillsMeta } from "../manageSkills/meta";
 import { META as manageSourceMeta } from "../manageSource/meta";
@@ -31,8 +29,6 @@ export const GENERATED_PLUGIN_METAS = [
   canvasMeta,
   chartMeta,
   editImagesMeta,
-  encoreDefineEncoreMeta,
-  encoreManageEncoreMeta,
   generateImageMeta,
   manageSkillsMeta,
   manageSourceMeta,