+++ src/plugins/_generated/server-bindings.ts
@@ -7,8 +7,6 @@ import { mcpEndpoint } from "../server-bindings-types";
 import accountingDef from "../accounting/definition";
 import canvasDef from "../canvas/definition";
 import chartDef from "../chart/definition";
-import encoreDefineEncoreDef from "../encore/defineEncoreDefinition";
-import encoreManageEncoreDef from "../encore/manageEncoreDefinition";
 import manageSkillsDef from "../manageSkills/definition";
 import manageSourceDef from "../manageSource/definition";
 import markdownDef from "../markdown/definition";
@@ -24,7 +22,6 @@ import spreadsheetDef from "../spreadsheet/definition";
 import { META as accountingMeta } from "../accounting/meta";
 import { META as canvasMeta } from "../canvas/meta";
 import { META as chartMeta } from "../chart/meta";
-import { META as encoreManageEncoreMeta } from "../encore/manageEncoreMeta";
 import { META as manageSkillsMeta } from "../manageSkills/meta";
 import { META as manageSourceMeta } from "../manageSource/meta";
 import { META as markdownMeta } from "../markdown/meta";
@@ -44,8 +41,6 @@ export const GENERATED_SERVER_BINDINGS: readonly ServerPluginBinding[] = [
   { def: accountingDef, endpoint: mcpEndpoint(accountingMeta) },
   { def: canvasDef, endpoint: mcpEndpoint(canvasMeta) },
   { def: chartDef, endpoint: mcpEndpoint(chartMeta) },
-  { def: encoreDefineEncoreDef, endpoint: mcpEndpoint(encoreManageEncoreMeta) },
-  { def: encoreManageEncoreDef, endpoint: mcpEndpoint(encoreManageEncoreMeta) },
   { def: manageSkillsDef, endpoint: mcpEndpoint(manageSkillsMeta) },
   { def: manageSourceDef, endpoint: mcpEndpoint(manageSourceMeta) },
   { def: markdownDef, endpoint: mcpEndpoint(markdownMeta) },